Change license string in pom.xml

+++ src/main/java/org/logicware/domain/geometry/Point.java
@@ -35,8 +35,8 @@ public class Point {
 
 	@Id
 	private String idp;
-	private double x;
-	private double y;
+	private int x;
+	private int y;
 
 	public Point() {
 	}
@@ -45,12 +45,12 @@ public class Point {
 		this.idp = idp;
 	}
 
-	public Point(double x, double y) {
+	public Point(int x, int y) {
 		this.x = x;
 		this.y = y;
 	}
 
-	public Point(String idp, double x, double y) {
+	public Point(String idp, int x, int y) {
 		this.idp = idp;
 		this.x = x;
 		this.y = y;
@@ -68,7 +68,7 @@ public class Point {
 		return x;
 	}
 
-	public void setX(double x) {
+	public void setX(int x) {
 		this.x = x;
 	}
 
@@ -76,7 +76,7 @@ public class Point {
 		return y;
 	}
 
-	public void setY(double y) {
+	public void setY(int y) {
 		this.y = y;
 	}
